chore: final cleanup
