+++ src/functions/schema-path.ts
@@ -25,13 +25,18 @@ export interface ISchemaPathOptions {
 }
 
 export const schemaPath: IFunction<ISchemaPathOptions> = (targetVal, opts, paths, otherValues) => {
+  if (!opts) {
+    // todo: nicer validation
+    throw new Error('No options provided');
+  }
+
   // The subsection of the targetValue which contains the schema for us to validate the good bit against
   const schemaObject = JSONPath({ path: opts!.schemaPath, json: targetVal });
-
   if (!Array.isArray(schemaObject) || schemaObject.length === 0) return;
 
+
   // The subsection of the targetVal which contains the good bit
-  const relevantItems = getLintTargets(targetVal, opts!.field); // todo: avoid this and validate
+  const relevantItems = getLintTargets(targetVal, opts.field);
 
   const results: IFunctionResult[] = [];
